Change collision particle for asteroid
Change collision particle (before particle is wrong use fire magix vfx)
But I can't use sub emitter. Because Iif I turn this on, appear the error, "sub-emitters must be children of the system that spawns them".
The cause of this problem is Unity Version, not me. If you can fix, please pull request.